+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
+## [1.4] 2025-10-12
+### Added
+- new extractor for audioblog.arteradiocom one item only
+### Changed
+- RadioFrance extractor with min. flatgreen/rfrance 2.5
 
 ## [1.3.3] 2025-06-08
 ### Fixed